Revised mirroring functions to have more accurate parameter types.

+++ src/utils.ts
@@ -384,10 +384,10 @@ export function checkLine(
 }
 
 /** Mirror a `Vector` over a horizontal line */
-export function mirrorHorizontal(origin: Vector, lineY: int): Vector {
+export function mirrorHorizontal(origin: Vector, lineY: float): Vector {
   return Vector(origin.X, -(origin.Y - lineY) + lineY);
 }
 /** Mirror a `Vector` over a vertical line */
-export function mirrorVertical(origin: Vector, lineX: int): Vector {
+export function mirrorVertical(origin: Vector, lineX: float): Vector {
   return Vector(-(origin.X - lineX) + lineX, origin.Y);
 }
